right shift the values by 8 in GAQ remap "B" table. data now fits in int8_t.
Originally committed as revision 14958 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/ac3dec_data.h
@@ -29,7 +29,7 @@ extern const uint8_t ff_eac3_hebap_tab[64];
 extern const uint8_t ff_eac3_bits_vs_hebap[20];
 extern const int16_t ff_eac3_gaq_remap_1[12];
 extern const int16_t ff_eac3_gaq_remap_2_4_a[9][2];
-extern const int16_t ff_eac3_gaq_remap_2_4_b[9][2];
+extern const int8_t  ff_eac3_gaq_remap_2_4_b[9][2];
 
 extern const int16_t (*ff_eac3_vq_hebap[8])[6];
 extern const uint8_t ff_eac3_frm_expstr[32][6];
